# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastSearch", "description": "Semantic search for fast.ai lectures.", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Index", "operationId": "index__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/feedback": {"post": {"summary": "Feedback", "description": "Insert search result feedback into feedback db.", "operationId": "feedback_api_feedback_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Feedback"}}}, "required": true}, "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/search": {"get": {"summary": "Search", "description": "Find lecture segments relevant to user query.", "operationId": "search_api_search_get", "parameters": [{"name": "query", "in": "query", "required": true, "schema": {"type": "string", "title": "Query"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Result"}, "title": "Response Search Api Search Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Feedback": {"properties": {"feedback": {"$ref": "#/components/schemas/Relevance"}, "query": {"type": "string", "title": "Query"}, "result_id": {"type": "string", "title": "Result Id"}}, "type": "object", "required": ["feedback", "query", "result_id"], "title": "Feedback"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Relevance": {"type": "integer", "enum": [1, -1], "title": "Relevance"}, "Result": {"properties": {"id": {"type": "integer", "title": "Id"}, "video_id": {"type": "string", "title": "Video Id"}, "title": {"type": "string", "title": "Title"}, "text": {"type": "string", "title": "Text"}, "start": {"type": "integer", "title": "Start"}, "thumbnail": {"type": "string", "title": "Thumbnail"}, "lesson": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Lesson"}, "forum": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Forum"}, "course": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Course"}}, "type": "object", "required": ["id", "video_id", "title", "text", "start", "thumbnail", "lesson", "forum", "course"], "title": "Result"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}